requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.4
django-admin-interface==0.29.3
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-social-oauth2==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jwcrypto==1.5.6
kombu==5.4.2
oauth2_provider==0.0
oauthlib==3.2.2
openai==1.59.6
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
PyTweet==1.3.6
redis==5.2.1
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
standard-imghdr==3.13.0
stripe==11.4.1
text-unidecode==1.3
tqdm==4.67.1
tweepy==4.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
